fix bugs
